ADD: mcommunity_groups to user model and populate it with a background job.
LDapable helper calls are expensive and time consuming. To try and reduce the impact of this, we are storing a user's mcommunity_groups (MCommunity groups that user is a member of and are open to the public api) in the user table.

We added an after_action to the OmniauthCallbacksController that calls the update_user_mcommunity_groups method. This method checks to see if the last time that user signed in is longer than 30 minutes. If it has been longer than 30 minutes, we instantiate the UpdateUserGroups background job. This job does the query (the bit that was taking so long)  to the LDAP server and creates an array of all of the groups the user currently belongs to, and then saves that to the user record to keep it fresh.

This is all in service of delegating user authorization to MCommunity groups so we don't have to deal with passwords and granting or revoking access.